requirements.txt
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.0
openpyxl>=3.1.0
matplotlib>=3.8.0
networkx>=3.2.0
pyarrow>=15.0.0
tabulate>=0.9.0
qstock>=1.3.8
tushare>=1.4.0
